requirements.txt
Django
Markdown
PyYAML
South
Werkzeug
django-admin-bootstrapped
django-bootstrap-toolkit
django-extensions
django-filter
djangorestframework
ipython
psycopg2
nose
